+++ lib/crypto.dart
@@ -4,4 +4,4 @@ library crypto;
 
 export 'src/crypto/formatting.dart';
 export 'src/crypto/keccak.dart';
-export 'src/crypto/secp256k1.dart';
+export 'src/crypto/secp256k1.dart' hide params;