+++ test/builder/data.dart
@@ -1,5 +1,5 @@
 const testCases = <String, String>{
-'''
+  '''
 [
     {
       "inputs": [
@@ -35,8 +35,7 @@ const testCases = <String, String>{
       "stateMutability": "pure",
       "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -77,7 +76,7 @@ class Retrieve3 {
   final bool var3;
 }
 ''',
-'''
+  '''
 [
     {
       "inputs": [],
@@ -128,8 +127,7 @@ class Retrieve3 {
       "stateMutability": "nonpayable",
       "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -190,7 +188,7 @@ class GiveMeHello {
   final BigInt num2;
 }
 ''',
-'''
+  '''
 [
     {
         "payable": true,
@@ -201,8 +199,7 @@ class GiveMeHello {
         "stateMutability":"payable",
         "type":"receive"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -219,7 +216,7 @@ class Contract extends _i1.GeneratedContract {
       : super(_i1.DeployedContract(_contractAbi, address), client, chainId);
 }
 ''',
-'''
+  '''
 [
     {
         "inputs": [],
@@ -234,8 +231,7 @@ class Contract extends _i1.GeneratedContract {
         "stateMutability": "view",
         "type": "function"
     }
-]''':
-'''
+]''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -268,7 +264,7 @@ class Contract extends _i1.GeneratedContract {
   }
 }
 ''',
-'''
+  '''
 {
     "abi": [],
     "devdoc": {
@@ -276,8 +272,7 @@ class Contract extends _i1.GeneratedContract {
         "methods": {},
         "version": 1
     }
-}''':
-'''
+}''': '''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;