+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.0
+
+- Add `EthPrivateKey.publicKey` getters
+- Fix `window.ethereum` always being non-null, even if no provider is available
+
 ## 2.1.4
 
 - Fix a generator crash for unexpected `devdoc` values